Convert Electronic Grader View to use New Model (#2970)
* Electronic Gradeable Status page

* Student list view

* Initial work on grading view

* student Information panel

* Removed last reference to old model from view

* Autograding configs use ints

* Misc work on details page

* from prev commit

* More work on details page

* Added method to check for version conflict in model

* Rough draft of details page

* Fixed a few parameter errors

* todo comment

* Details view handles non-team students

* more consistent name

* student information panel uses new model for select

* Updated helper function to be more accurate

Since the getorCreateTaGradedGradeable function exists, we may not have any grades despite the object being non-null

* Update display settings for 'grade' button

* Re-added 'new team' button for user

* More consistent with previous behavior

* Reorganized flow of teamless users

* Fix student information panel

* Removed old code

* removed old comment

* Fixed ordering of users in team

+++ site/app/models/gradeable/AutogradingTestcase.php
@@ -27,7 +27,7 @@ class AutogradingTestcase extends AbstractModel {
     protected $name = "";
     /** @property @var string TODO: The command to run? */
     protected $details = "";
-    /** @property @var float Number of points this testcase is worth */
+    /** @property @var int Number of points this testcase is worth */
     protected $points = 0;
     /** @property @var bool If this testcase is extra credit */
     protected $extra_credit = false;
@@ -49,7 +49,7 @@ class AutogradingTestcase extends AbstractModel {
 
         $this->name = Utils::prepareHtmlString($testcase['title'] ?? '');
         $this->details = $testcase['details'] ?? '';
-        $this->points = floatval($testcase['points'] ?? 0);
+        $this->points = intval($testcase['points'] ?? 0);
         $this->extra_credit = ($testcase['extra_credit'] ?? false) === true;
         $this->hidden = ($testcase['hidden'] ?? false) === true;
         $this->view_testcase_message = ($testcase['view_testcase_message'] ?? true) === true;
@@ -57,7 +57,7 @@ class AutogradingTestcase extends AbstractModel {
 
     /**
      * Gets the number of points this testcase is worth if $this->hidden is false, otherwise 0
-     * @return float
+     * @return int
      */
     public function getNonHiddenPoints() {
         return (!$this->isHidden()) ? $this->points : 0;
@@ -66,7 +66,7 @@ class AutogradingTestcase extends AbstractModel {
     /**
      * Gets the number of points this testcase is worth if $this->hidden is false and $this->extra_credit is false,
      *  otherwise 0
-     * @return float
+     * @return int
      */
     public function getNonHiddenNonExtraCreditPoints() {
         return (!$this->isHidden() && !$this->isExtraCredit()) ? $this->points : 0;
